+++ src/Field/Field/StaticMultiSearchSelect/StaticMultiSearchSelectField.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace srag\RequiredData\Field\Field\StaticMultiSearchSelect;
+
+use srag\RequiredData\Field\Field\MultiSearchSelect\MultiSearchSelectField;
+
+/**
+ * Class StaticMultiSearchSelectField
+ *
+ * @package srag\RequiredData\Field\Field\StaticMultiSearchSelect
+ *
+ * @author  studer + raimann ag - Team Custom 1 <support-custom1@studer-raimann.ch>
+ */
+abstract class StaticMultiSearchSelectField extends MultiSearchSelectField
+{
+
+    /**
+     * @var string
+     *
+     * @abstract
+     */
+    const TABLE_NAME_SUFFIX = "";
+}