+++ src/Field/Field/StaticMultiSearchSelect/StaticMultiSearchSelectFillField.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace srag\RequiredData\Field\Field\StaticMultiSearchSelect;
+
+use srag\RequiredData\Field\Field\MultiSearchSelect\MultiSearchSelectFillField;
+
+/**
+ * Class StaticMultiSearchSelectFillField
+ *
+ * @package srag\RequiredData\Field\Field\StaticMultiSearchSelect
+ *
+ * @author  studer + raimann ag - Team Custom 1 <support-custom1@studer-raimann.ch>
+ */
+abstract class StaticMultiSearchSelectFillField extends MultiSearchSelectFillField
+{
+
+    /**
+     * @var StaticMultiSearchSelectField
+     */
+    protected $field;
+
+
+    /**
+     * @inheritDoc
+     */
+    public function __construct(StaticMultiSearchSelectField $field)
+    {
+        parent::__construct($field);
+    }
+}